backend/pyproject.toml
[project]
name = "m3-mix-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pymc-marketing>=0.9.0,<0.20.0",
    "scipy>=1.13.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "groq>=0.9.0",
    "python-multipart>=0.0.9",
    "supabase>=2.7.0",
    "httpx>=0.27.0",
    "pydantic>=2.7.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]
